Restructure the source files
