+++ lib/compress/zstd_compress_internal.h
@@ -82,25 +82,6 @@ typedef struct {
 } ZSTD_entropyCTables_t;
 
 typedef struct {
-    symbolEncodingType_e hType;
-    BYTE hufDesBuffer[500];
-    size_t hufDesSize;
-} ZSTD_hufCTablesMetadata_t;
-
-typedef struct {
-    symbolEncodingType_e llType;
-    symbolEncodingType_e ofType;
-    symbolEncodingType_e mlType;
-    BYTE fseTablesBuffer[500];
-    size_t fseTablesSize;
-} ZSTD_fseCTablesMetadata_t;
-
-typedef struct {
-    ZSTD_hufCTablesMetadata_t hufMetadata;
-    ZSTD_fseCTablesMetadata_t fseMetadata;
-} ZSTD_entropyCTablesMetadata_t;
-
-typedef struct {
     U32 off;
     U32 len;
 } ZSTD_match_t;
